Keep unknown PPE 400 sections and carry declared names
Four points were left open when the sectioned container landed, and one
of them was a real defect rather than a missing convenience.

The content identity covered every section, so a file written by a newer
compiler with any additional optional section no longer matched its own
digest and this loader refused it. That contradicts the rule that an
optional section may be skipped. The identity now covers the six sections
that make up the program, which keeps it valid across repacking, stripped
debug data and future additions. Unknown optional sections are also kept
verbatim through load and save instead of being dropped, so rewriting a
file no longer discards data this version does not understand. An unknown
compression code is now treated as malformed: compression belongs to the
container version, not to a section's meaning.

Debug data carried variable names only, which left the decompiler
inventing TYPE001 and FIELD001 for programs that had perfectly good names.
It now also carries record, field and enum names, and the decompiler emits
them. The section stays structurally complete when a name is missing, so a
program built without names still loads. Stripping the names changes
neither behaviour nor the content identity.

Record field bounds were still 16 bit in memory although the format stores
32, and the parser rejected anything above 65535. Both now match the
format. The compiler measured an oversized program in the byte count of
the dropped word encoding; it now measures the CODE section it actually
produces. The legacy target keeps its old limit and diagnostic.

Adopting the stored variable names in the decompiler was tried and
reverted: a function and its result share a name in PPL, so the recovered
source no longer compiled. Type, field and enum names have their own
namespaces and are unaffected.

S7 is covered by a new test that exercises the S1-S6 contracts together in
one program - a record with host, dynamic and enum fields, a short circuit
operator, VAR write-back, a nominal cast, Unicode lengths and an ON ERROR
handler - written as a real file in both compression modes, executed,
decompiled and rebuilt with the same output.

Verified with the five PPL packages in separate English and German
processes: 2907 passed, 0 failed, 6 ignored in each. The all-target check
of those packages and the engine without default features pass.

Still open, as listed in the release plan: no source positions and no
separate debug file, read_file reads before the size check, Zstd is built
and tested on Linux only, and the loader has not been fuzzed.

+++ crates/icy_board_engine/src/vm/host_defaults.rs
@@ -248,8 +248,8 @@ mod tests {
                     RecordField {
                         dim: rank,
                         vector_size: 1,
-                        matrix_size: u16::from(rank >= 2),
-                        cube_size: u16::from(rank >= 3),
+                        matrix_size: usize::from(rank >= 2),
+                        cube_size: usize::from(rank >= 3),
                         ..RecordField::scalar(VariableType::UserData(102))
                     },
                 ],
